A Link Between Worlds
An expanded port of stdlib's DoublyLinkedList.  More changes than
last time, mostly to buff up the individual Nodes with more ability
than none at all.

I realized when I got to the end that stdlib's version trusts the
Nodes with nothing because... it doesn't null out links when nodes
are removed.  I'm sorry but that is not how I was raised.